requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.12.0
streamlit==1.29.0
pandas==2.1.4
openpyxl==3.1.2
chromadb==0.4.22
faiss-cpu==1.7.4
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.24.3 